requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
couchbase==3.0.6
cryptography==3.2.1
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
importlib-metadata==2.0.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyrsistent==0.17.3
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
